requirements.txt
wheel
requests
python-kasa
simplemoer
